requirements.txt
apify_client
reportlab
python-dotenv
meta_ai_api
requests
